Linked CSS & JS file and fixed things and necessary changes
One major thing that was missing was the linking of the CSS file and the JS file. I noticed when you type in any of the input fields it was displaying the previous text like 'First Name' with the name we input. So, I fixed it by adding a placeholder instead of Label.